requirements.txt
llama-cpp-python
tqdm
requests